Добавление записи в бд

+++ My project/Assets/Scripts/Core/DatabaseProvider.cs
@@ -70,7 +70,7 @@ namespace Core
 
         public void AddWorkoutRecord(long userId, long coachId, DateTime dateTime, string coachSum)
         {
-            dataBaseConnector.AddWorkoutRecord(userId, coachId, dateTime, int.Parse(coachSum));
+            dataBaseConnector.AddWorkoutRecord((int)userId, (int)coachId, dateTime, int.Parse(coachSum));
         }
     }
 }
